back/requirements.txt
airsim==1.8.1
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.57.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-seed==0.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Faker==8.1.1
idna==3.4
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
numpy==1.26.2
oauthlib==3.2.2
opencv-contrib-python==4.8.1.78
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
requests==2.31.0
requests-oauthlib==1.3.1
six==1.15.0
sqlparse==0.4.4
text-unidecode==1.3
tornado==4.5.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
